Sujet 2/requirements.txt
pip
pillow
